Chore[tests]: Move to tests/integration
