+++ tests/integration/mex.test.js
@@ -1,13 +1,13 @@
-const { salesTax } = require('./getTax');
+const { salesTax } = require('../../getTax');
 
 describe('Mexico', () => {
   describe('IVA', () => {
     test('Normal', () => {
-      const o = {
+      const q = {
         taxType: 'IVA'
       };
 
-      expect(salesTax('mx', o)).toEqual(
+      expect(salesTax('mx', q)).toEqual(
         expect.objectContaining({
           result: 0.16
         })
@@ -15,13 +15,13 @@ describe('Mexico', () => {
     });
 
     test('Pharmacy Drug', () => {
-      const o = {
+      const q = {
         taxType: 'IVA',
         category: 'DRUG',
         area: 'PHARMACY'
       };
 
-      expect(salesTax('mx', o)).toEqual(
+      expect(salesTax('mx', q)).toEqual(
         expect.objectContaining({
           result: 0
         })
@@ -30,11 +30,11 @@ describe('Mexico', () => {
   });
 
   test('RET_IVA', () => {
-    const o = {
+    const q = {
       taxType: 'RET_IVA'
     };
 
-    expect(salesTax('mx', o)).toEqual(
+    expect(salesTax('mx', q)).toEqual(
       expect.objectContaining({
         result: 0.167
       })
@@ -42,11 +42,11 @@ describe('Mexico', () => {
   });
 
   test('RET_ISR', () => {
-    const o = {
+    const q = {
       taxType: 'RET_ISR'
     };
 
-    expect(salesTax('mx', o)).toEqual(
+    expect(salesTax('mx', q)).toEqual(
       expect.objectContaining({
         result: 0.1
       })
